+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - SelectableAudioMenu component
 - SubmitableAudioMenu component
 - SelectWhenHighlight component
+- MenuData Scriptable Object
 - AbstractMenu component
 
 ### Changed